refactor: align with packages implementation
