+++ spog/ui/src/pages/advisory/mod.rs
@@ -5,13 +5,9 @@ use spog_model::prelude::*;
 use yew::prelude::*;
 use yew_more_hooks::hooks::{UseAsyncHandleDeps, UseAsyncState};
 
-use crate::{
-    components::{
-        advisory::{AdvisoryResult, AdvisorySearch},
-        common::PageHeading,
-        simple_pagination::*,
-    },
-    hooks::use_pagination_state::*,
+use crate::components::{
+    advisory::{AdvisoryResult, AdvisorySearch},
+    common::PageHeading,
 };
 
 // FIXME: use a different API and representation
@@ -35,30 +31,15 @@ pub fn advisory(props: &AdvisoryProps) -> Html {
     };
     let query = props.query.clone();
 
-    // Pagination
-    let total = search.data().and_then(|d| d.total);
-    let pagination_state = use_pagination_state(|| UsePaginationStateArgs {
-        initial_items_per_page: 10,
-    });
-
     html!(
         <>
             <PageHeading subtitle="Search security advisories">{"Advisories"}</PageHeading>
 
             // We need to set the main section to fill, as we have a footer section
             <PageSection variant={PageSectionVariant::Default}>
-                <AdvisorySearch {callback} {query} pagination={pagination_state.clone()}/>
-
-                <AdvisoryResult state={(*search).clone()} />
-
-                <SimplePagination
-                    position={PaginationPosition::Bottom}
-                    total_items={total}
-                    page={pagination_state.page}
-                    per_page={pagination_state.per_page}
-                    on_page_change={pagination_state.on_page_change}
-                    on_per_page_change={pagination_state.on_per_page_change}
-                />
+                <AdvisorySearch {callback} {query}>
+                    <AdvisoryResult state={(*search).clone()} />
+                </AdvisorySearch>
             </PageSection>
         </>
     )
